Removed a number of glob imports
Glob imports meaning `from x import *`. I don't think it's called that
in python.

Glob imports are near-universally bad, as not only do they make
statically detecting undefined variables impossible, they're also just a
bad development practice to begin with.

Quite a few remain in gui/colors/, but the imports in there are a
complete and utter mess so I'm planning to refactor it around.