Only auto-enable the frame on the first activation
Fixes #79.

The frame tool now keeps track of whether or not it's being activated
for the first time. Previously, once it was switched back to after a
panning/rotating/etc operation, the frame was automatically enabled
regardless of whether it was before, which was very frustrating.

However, if the frame is disabled and the user manually switches to the
frame tool, it'll still be re-enabled. I'm not sure why this happens,
but it is actually a desirable effect, as to the user functionality is
near-identical but with one less bug.