Popup Menu Fix
Address the Problem in issue #94 where using a two finger
tap/press(right-click) on Multi-Touch touchpads and clickpads causes the toolbar
menu immediately unfocus. This was cause by the gdk.CURRENT_TIME. The
fix was to use the gtk.get_current_event_time() which was used in
the popup menu for the layers window.

This should fix #94