Changed history-adding logic to only compare names
Previously it used BrushInfo.matches to compare same-ness, which meant
that if the user changed a small setting such as brush size, it would
count as an entirely separate brush and have it's own entry in the brush
history. This was undesirable, and simplifying the code to only compare
the brush names makes the history less likely to be polluted with small
changes.