Changed Zoom Tool functionality
This is for consistency (with the Rotate Tool) and ease of use.
Previously, it kept centred on the mouse at all times, which proved
awkward as it caused you to pan up/down as you zoomed. It now always
keeps centred on the middle of the screen, in the same vein as the
Rotate Tool.