Tentative build docs for Windows
MSYS2 is comprehensible to my tiny, Linux-polluted brain, and it seems to
be a good way of getting Windows stuff built and tested. Which is nice.

Add some tentative instructions for building and testing on Windows.
Status: it compiles. And that's it, don't get your hopes up yet. MyPaint
doesn't actually run yet (there will be tons of Python-side things to clean
up first), but that can be worked on now!

Fixes #48.