Tweak the Windows build instructions a bit
json-c is now available via msys2, which simplifies things a lot. No longer tell users to build it themselves, and add it to the list of deps to grab from the msys2 repositories.

Looks like merely having msys2 scons and its python2 available don't actually interfere with the build, as I feared. We do still need scons-local though, for whatever reason. Update the docs to reflect that, and also to keep scons-local stuff outside the mypaint tree.