Edit Composting Mode Tool Tips.
I've edited the tooltips strings in ```/lib/layer.py``` for the Composting Mode 
descriptions. These explanations are a bit more indepth to help artists understand
what they do. Plus the tooltip for the Destination compositing layer didn't specify that 
they can be used for masking out or in object.

TODO: Probably search for a way to insert text wrapping to the tooltips instead of 
manually typing the newlines in the strings.