Introduce SAVE_FORMAT_PNGAUTO
SAVE_FORMAT_PNGAUTO is a special PNG save format for when the user does
not explicitly choose to save a solid or transparent PNG (i.e. when the
user does a 'save as' with a filename ending in '.png' without touching
the 'Format to save as' dropdown menu, or when the user tries to 'save'
a PNG after previously doing a 'save as').

If the background is visible, a solid PNG will be saved. Otherwise, a
transparent PNG will be saved.

Fixes the other half of github issue 165.