Layer loading: fix initial selection logic
The old code was producing paths for the initial selection which didn't
exist in the document due to using only a simple count.
This would have been only apparent when the loaded document had no layers
marked as selected *and* also had nested layers.

Closes #196. Thanks for the report and investigation, @vurentjie!

Taking the opportunity to fix up the logic so that it selects either a) the
topmost layer that's marked with selected="true", or b) the topmost
immediate child layer of the root, which is always something that exists.
I think using the highest-up layer when we have multiple choices will be
more convenient for users.