Fix segfault when dragging tabs in gtk 3.14.8
Quick and dirty fix for a segfault seen when dragging tabs to make a new
window in the newest stable release of GTK. The fault is ours; we're
playing fast and loose with what GTK expected in a handler with defined
job. We're still cheating, but in a way that doesn't freak GTK out.

Fixes mypaint/mypaint#194.