Cosmetic: frame size dialog: headers, spacing
Bump up the spacing in the frame size dialog, and give it some headers.
Hopefully this will make it less confusing.

Addresses mypaint/mypaint#217.

The DPI setting is saved into OpenRaster files, so we can't hide it
when the units are pixels even when it makes no sense. It's still a
nominal, saved resolution.