Add a basic Inking tool
Add a tool for making smooth, controllable curves. The idea is that it's
a bit like freehand mode, but allows the last curve you drew to be tweaked
just after you drew it.

It works in two phases:

1. Draw a line to record a sequence of control points
2. Tweak the control points
3. Accept or reject the tweaked curve with onscreen buttons,
   or just start a fresh one

Curve rendering uses a basic Catmull-Rom spline, but we can change it to
something fancier later.  Re-rendering happens in an idle routine to allow
the curve points to be adjusted responsively.

Closes mypaint/mypaint#13