Ink tool: workaround for dropouts with evdev
Workaround for evdev drivers sending zero pressure when the pressure has
not changed since the last event. Yes, it's this bug again!

This affects lots of cheap, popular Monoprice, Genius and Huion tablets,
and probably other brands. If your tablet has "UC-Logic" in the name X11
presents, it probably needs this.

Closes #223 (tested with Wacom Intos5 + evdev driver)

Much of this inktool stuff has been sitting around in a local branch of
mine for ages, since before the fix for #29.