Model's pending changes: more subtlety
Code can now ask for sources of pending changes to the model to synchronize
their changes without having to flush them to the model.

Change the @event name to reflect that it's become a (potentially
bidirectional) synchronization point.

Update modes' checkpoint() interface to handle the new flag, and update its
docs.