Reorder path detection
Fix a refusal to run when invoked as "mypaint" rather than "./mypaint" from
the root of the source tree if there is a normal install elsewhere in the
user's $PATH.

Closes mypaint/mypaint#228.

You'll only see this fix take effect when you update that other
installation. Luckily this bug is a super-rare thing that'll probably only
bite developers and testers.