Refactor: move layer tree model
Move the layer TreeView to a side module away from the panel one to allow
independent testing of it.

Keep the current broken way of handling drags, for now.  This will be
replaced with low-level GTK DnD code because the annotations for
gtk_tree_set_row_drag_data() and its get() cousin have historically been
very broken, and because our use of the high level interface just stopped
working one day. I basically give up trying to use the high-level GTK
listview DnD protocol at this point; we have too many versions of GTK to
support for the next release.

This drops the ability to move into/under a non-group layer using the shift
key. It was cool functionality, but it's in the way when doing this kind of
rewrite.

Addresses mypaint/mypaint#148.