Layers panel: do our own drag-and-drop handling
Rip out the old dtd implementation (using the high-level GtkTreeDragSource
and GtkTreeDragDest interfaces), and replace it with low-level dnd handlers
on the corresponding view.

Relying on these interfaces was proving very unreliable from one version of
GTK to another for reasons I can't quite fathom. I suspect it's to do with
GtkSelectionData access and typelib annotation of its manipulation funcs,
but I'm not certain.

The low-level dnd interface shared by everything should be more stable and
reasonably permanent, and has the advantage that upstream actually test it.

Fixes mypaint/mypaint#148.