Remove overrides for gtk-dnd-drag-threshold
Overriding this was causing issues with sensitivity in the layers menu,
and doing it twice resulted in the dnd sensitivity being dependent on
panel initialization order. In other words, different people were getting
different results :/

Addresses mypaint/mypaint#240.

This may make brush dragging, or clicking in the brush lists, seem
oversensitive to some people (that's the reason the override that the
override was there at all, IIRC). Testing would be appreciated.

The primary way to adjust this right now is by configuring your settings
manager with a more comfortable drag threshold. On GNOME or Ubuntu Unity,
try

gsettings set org.gnome.settings-daemon.peripherals.mouse drag-threshold 16

if the default of 8 pixels is bothersome.