[feature] Make auto-saves when changes settle
Write an autosave folder out to the cache after a short period of
inactivity following a change. The file trees now written to
<cache>/mypaint/doc.XXXXXX/autosave resemble OpenRaster files in structure,
and should have useful guarantees of internal consistency.

The <cache> folder used for all of this is the one GLib tells us (see
the docs for g_get_user_cache_dir()). On POSIX systems we use
$XDG_CACHE_HOME, normally $HOME/.cache. On Windows 7, a typical path is
%LOCALAPPDATA%\Microsoft\Windows\Temporary Internet Files, which has the
disadvantage of being a hidden location (but then, so is Linux's with the
dot-name and everything).

The idea is to give people a way of recovering files when mypaint crashes
hard when changes haven't been saved to an output file. If it does, and
you're lucky enough to have a good autosave backup, you should be able to
zip it into a new .ora file and recover data that way.

Partly addresses mypaint/mypaint#90.
For a full solution, we'll need a real auto-recovery dialog.

Consistency
-----------

If you see a file named stack.xml in an autosave "doc.*" folder, then
everything it indexes can be expected to exist, and it should even have a
valid thumbnail. Neither stack.xml nor the files it indexes will be
partially written if they exist at all, assuming the file system hasn't
failed.

This code doesn't attempt to guarantee consistency when *filesystems*
crash, although some measure of security could be gained there by using
flushes and fsync() calls. There are however good performance reasons not
to do that, and MyPaint already tries to provide good data robustness when
saving normally.

Performance
-----------

The bulk of the work is found in saving big layer images and their
strokemaps. To keep auto-saving responsive, this code queues all the
writes in lots of separate idletask chunks, at a very low priority.
Hopefully this means that autosaves don't interfere too much with
high-priority foreground stuff like event capture and brush engine
performance.

Layers now have an autosave_dirty flag which must be set when their file
data should be saved. Autosave passes don't write more then they need to.

I'd like to use threads for this, but the interactive performance when I
tried that seemed *remarkably* bad. The writes just choke off all other
activity, perhaps for GIL-related reasons? The fact that you can't lower
the priority of Python threads doesn't help much either.