gui.mode: Smooth scrolling, no rotation
Stop scrolling doing rotation. This was *incredibly* disconcerting with
touchpads configured to do bidirectional scrolling. This is not an uncommon
configuration these days.

Fixes mypaint/mypaint#204.

Scrolling is now smooth for devices which support it. Hold down Shift to
make smooth scrolls discrete again (i.e. to make it work like before).

Hold down Alt or Ctrl to pan instead of zooming.

If we want to support scroll rotation again, suggest Ctrl+Scroll. But it's
absent for now because I don't really see a nice way of doing it.

Zooming is still the default behaviour when unconfigured, but MyPaint now
supports panning rather than zooming as the unmodified action for a
device's scroll events. This can be configured in the device preferences,
and can be different between devices.