nav-only devices: let btn1 start drags like btn2
If a hardware ("slave") device is configured to be a navigation-only
device, then treat its button-1 as a supplementary button-2 for the purposes
of initiating temporary mode switches based on pointer clicks and drags.
By default, button-2 does navigation (viewport panning, rotation, and
zooming).

This allows single-finger drags with an appropriately configured
touchscreen, in lieu of real touchscreen support. The AllowedUsage setting
for a device (navigation-only vs. allowed-for-painting) is also how I want
to distinguish between the finger-painting and single-finger-panning styles
when we *do* implement real touchscreen support, so let's build in a bit of
that behaviour now.

This is roughly the way the space bar drags work, so it's not that weird a
concept. Really!

Okay, it's slightly experimental, and let's rip this out if it breaks
stuff. However, it does have a nice feel on my XPS13.

Closes mypaint/mypaint#246.