gui.picker: set event mask when starting the grab
Set up the event mask in _start_grab(). Closes mypaint/mypaint#324.

Setting up both the model and the view by feeding in a single Application
god-object isn't great MVP, but it's a bit close to beta now to change
that. Just add a note about it for now.

One day, I'd quite like to rejig the way the painting canvas is architected
to be more MVP-like too, so it's worth nailing down how the pattern works
in a Gtk.Builder world. Future me, please excuse past me's early
experiments here.