Use a MenuButton for toolbar brush/color menus
Using the old GtkMenuToolButton hack for the toolnbar's brush and color
menus looked confusing.  Replace them with a GtkToolItem containing a
GtkMenuButton, which has been around since GTK 3.6.

There's no such thing as a MenuButtonToolItem in GTK, so still need to
implement our own along with the action architecture to bring it up.

Closes mypaint/mypaint#143.