Allow smooth scrolling to be turned off globally
Smooth scrolling was proving annoying to our testers. However the new
behaviour can only truly be turned off by unsetting the event mask and
making GDK deliver discrete/stepwise scroll events for all devices.

Add a checkbox to the View tab of the prefs window that turns off smooth
scrolling for all TDWs after a restart.

Also fix the layout of some labels in the dialog on that tab.

Closes mypaint/mypaint#328.