Fix exception with weird interleaved input streams
Fixes an OverflowError exception seemingly caused by the pressure and tilt
interpolator we use in certain input cases: just clamp all pressure and
tilt data before feeding it to the C/swig brush engine wrappers.

The problem can arise in cases where zero-tilt data interleaves with tilted
data, for example when knocking a mouse bound to the same pointer as a
tablet you're currently drawing with.  It doesn't happen with all
combinations of input device though, for unclear reasons (lacking axes? Bad
data on some axes?)

The interpolation method we use seems to be responsible, but for now we can
just work around the resultant bad data from it.

Closes mypaint/mypaint#344.