symmetry: draw line using an alpha read from prefs
Draw the symmetry line and its shadow to the same cairo group, which is
then rendered with alpha. The base alpha can be set by the user in the
on-disk prefs (UI TBA in next few commits).

The alpha ultimately used while editing the line varies with the cursor's
distance from the axis, making it look more active and allowing the user to
preview their choice for the value of the alpha setting. While the cursor
is far away, the alpha is whatever the user will choose in the panel UI.
When the cursor is directly on top of the axis, the alpha is 100%.

Addresses mypaint/mypaint#295.