symmetry: UI to set the axis's alpha
The axis of symmetry can be a bit visually intrusive while painting.

This commit lets the user set an alpha value between 0% and 100% for the
symmetry axis line while painting: it adds a slider in the options panel
for the Symmetry Axis adjust tool. Setting the alpha to zero will make the
axis invisible. The effect is immediately previewable in the axis edit mode
you set the alpha from, so it should be reasonably obvious and
discoverable.

Closes mypaint/mypaint#295.