lib.errors: add OOM wrapper exception class
Code which is capable of detecting possible OOM conditions should log
whatever condition or exception it has to hand, then raise a
lib.errors.AllocationError in its stead with a translated user-facing
message string.

Code expecting to receive one of these and provide hints to the user should
also catch MemoryError too, to be on the safe side. AllocationError is for
when you suspect GI-wrapped code failed due to OOM and spat out a NULL or a
different exception.