pixbufsurface: raise AllocationError appropriately
GdkPixbuf.Pixbuf.new() currently loves to raise TypeError when it fails to
allocate the pixbuf, but it could basically be any sort of error return
depending on how the annotations change. Catch the obvious oopsies and give
the user a "hey, couldn't allocate $IMPORTANT_THING" message.

Addresses mypaint/mypaint#62.

Also update the docstrings and comments.