workspace: add method for revealing a panel
As opposed to adding/removing the panel widget, which reorders the UI.
Sometime's it's nice to just bring up a specific panel with a keypress.

Revealing means showing the widget's enclosing ToolStack, adding it
first if it doesn't exist.

Refactor getting a widget's enclosing toolstack now that's used in two
places.

Fix minor typo.