Add namespace for MyPaint's OpenRaster extensions
MyPaint needs to store a number of additional fields in OpenRaster files
for its own benefit. The natural way of doing this in XML is to use a
namespace, so I've picked "http://mypaint.org/ns/openraster" - which isn't
an addressable resource yet, but it a) doesn't have to be, and b) could be
one day if it's useful to have something there.

Okay, OpenRaster doesn't even have its own xmlns yet, but no reason not to
be ahead of the curve.

Addresses mypaint/mypaint#18.