OpenRaster: save+load whether the frame is active
Save a custom xsd:bool declaring whether the frame is on or off in an XML
attribute in OpenRaster files and autosave oradirs. It's a distinct
concept from image size, and shouldn't be confused with it.

Update the frame-enabled state from the saved value when loading OpenRaster
files or oradirs files. The default is no frame (infinite canvas).

Closes mypaint/mypaint#18.