Extend support for reading old strokemap
Egg-on-face moment: I am a dope, and I can't read specification documents.
Semver practice is to deprecate in one MINOR version, and retain support
until the next MINOR version, not patch version.

Reading the old strokemap attribute and file format is now due to be
discontinued in v1.3.0, not v1.2.1. Oops.