Layers list: don't always scroll on content change
Only scroll the list to the current layer on a layer content change
notification if the layer that's been changed is the current layer.

This change allows the user to scroll in the layers panel so that the
current layer is no longer visible, and then toggle lots of layer
visibilities in that area of the list. Before, the list would scroll to the
current layer after the 1st click.

Closes mypaint/mypaint#308.

Drawing on the current layer will still scroll the layers panel so that
the layer being drawn on is shown in the list. This is intentional, and
was the primary use case for this bit of UX design.