Layer Solo: respect the background visibility flag
Reverse an earlier decision to always show the background in layer solo
mode. It's more useful to have two distinct renderings for a solo-mode
layer, controlled by the Show Background toggle:

* When the background is shown, the current layer affects it directly.
  This makes it obvious what the layer currently does.

* When the background is hidden, the current layer is rendered in Normal
  mode over the background checks.
  This makes it easier to edit the "real" data in a layer.

When the current layer is being blinked on a layer change, the
show-background state is temporarily inverted to make the blink look more
distinct.

Closes mypaint/mypaint#233.