Version numbering scheme update
Move away from date-based version numbering for the first series of
beta releases, largely for aesthetic reasons but also because I'll
be tagging release points in git for the betas, making "git describe"
a useful tool. Version numbers in builds will resemble:

    1.2.0[-beta.0[.3]][+git[export].1234abc]
    |     |     |  |    |
    |     |     |  |    Build/export info (not part of tarball)
    |     |     |  |
    |     |     |  3rd commit after the "v1.2.0-beta.0" tag.
    |     |     |
    |     |     Beta release 0; corresponds to a release tag.
    |     |
    |     Pre-release marker: beta releases.
    |
    Version this is a pre-release of.

It is now very important to have tags available in the git checkout,
and for them to match up with the base revision in the code, so this
is now enforced. Building from an exported source tarball will still
work without git however, and this style of build will form the basis
for binary prereleases in the form of Windows installers released
through github, and Debian packages released via the Ubuntu PPA.

* lib.meta: update docs.
* lib.meta: bump to v1.2.0-beta.0.
* lib.meta: when called as __main__, write the entire hybrid
  metainfo script on stdout.
* lib.meta: enforce git tagging requirements when extended version
  information is requested and this isn't an unpacked source tarball.
* SConscript: update for the new calling convention.
* release.sh: update for the new calling convention.
* release.sh: remove reminder to tag - that gets enforced now.