Auto-save recovery dialog: exception fixes
Internal MyPaint paths are unicode objects, but they also have to be stored
in a treestore for the autosave recovery dialog. Convert the column values
to and from utf8-encoded byte strings to avoid exceptions when clicking
around in the autosave recovery list.

Fixes mypaint/mypaint#374 for me on Win7 64bit Professional with various
flavours of non-ASCII %USERPROFILE% paths.