Fix mypaintlib import order screwing up i18n
If mypaintlib is imported too early, i.e. before the i18n setup, it can
mess up what the GLib get_*_dir functions return. This probably indicates
that it or something it uses is doing something wrong, but it's easier to
fix this by doing the checker import later than by chasing down the source
of the error.