Make expanduser_unicode do nothing on Win32
The encoding used by sys.getfilesystemencoding() isn't roundtrippable,
which breaks scrap saving. The "~" convention has no real meaning on
Windows, and there are better defaults for things like the scrap prefix on
that platform, so just have the func return its input coerced to unicode.

All paths that MyPaint stores in RAM or saves to the config should be using
Unicode anyway.