brush dnd: make it less weird
When dragging a brush from a source group into a separate target group
where it already exists, don't do anything.

The old behaviour was weird, and I keep having to explain it to people
which is generally a bad sign. Before this commit, the brush would be
repositioned within the target group *and* forcibly removed from its source
group causing lots of updates and all the tiles shifting around.
Additionally, having different behaviour on a second attempt to drag an
item from one group to another was, er, "unique".

It was like that because this was the only way of removing a brush from a
group until now. The context menu fixes that now.