filehandling: improve recent-files filtering
Add a custom file-filter function so that only files we're reasonably sure
MyPaint will be able to load as editable main documents show up in the
recent files menu. The test is now requires app association, a recognized
file extension, and existence of the file.

When updating the internal var that Open Most Recent (F3) uses, use the
same logic as the filter function.

Closes mypaint/mypaint#267.