Layer panel: flip eye+lock and type icons around
Putting the eye and lock icons on the left (for ltr languages) is closer to
other art apps.

I had my doubts about doing this during string freeze, but I guess for a
sensible layout, you don't need fancy multi-cell columns after all. So no
string changes needed :)

Second advantage: this makes Adwaita's new (and horrid) mystery meat
scrollbars break things much less. There's almost no need to click on the
type icon for a layer, so it doesn't really matter when the scrollbar fades
in above it. Before we had the lock icons on that side, and it was
difficult to click on them when there were enough layers in the stack to
require a scrollbar.

Closes mypaint/mypaint#144.