Device monitor: fix missing notifications
Fix: Windows users can now configure devices individually.

It seems that Windows GDK 3.16 doesn't send device-added or device-removed
signals when new devices are plugged in or unplugged. This meant that the
Devices tab in Preferences wasn't being updated on Windows builds because
it was relying on this mechanism for detecting changes to the list of known
devices.

Improve logging of new devices, and downgrade the messages for the
GdkDeviceManager signals to debug level now that we're not relying on them.

Addresses an oddity noticed as part of mypaint/mypaint#401.