Profiler improvements
Now any system can run the profiler nicely, not just Linux-based ones
with gprof2dot.py and dot installed. The output without the tools is
more minimal, of course, but the .pstat files can be copied elsewhere
for analysis and comparison.

Document profiler usage (in a new DEBUGGING document).

Use a special managed tempdir for profiler output. It's cleaned up at
the end of the session to keep things tidy.