strokemap: API for finishing pending tile work
Add an internal method for processing just specified queued tile work, if
that's possible. It's only possible if every chunk of work in the idletask
work queue supports being partially forced to completion for specified
tiles: if the queue involves big global changes like translating/moving
tiles, we'll have to use the old way still.

This is intended to speed up stroke picking and blinking.