strokemap: speed up stroke render (real tile obj)
It's not unusual for the tiles in the strokemap to be all ones, especially
for big strokes of the kind that cause problems in mypaint/mypaint#8.

Those can be stored in RAM slightly more efficiently (though zlib does a
decent compression job), but there's a nice rendering speed gain to be had
for strokes where all-ones tiles predominate over partly-filled times. From
my measurements on a completely biased sample of one big image with big
wide strokes: ~ 2.1x speed and no loss or much gain on an image with
comparable big narrow strokes.

Should help some more with the user experience in mypaint/mypaint#8, but I
suspect the lingering-stroke part of that is already fixed by preceding
commits.