Restore old color picker behaviour (hover to pick)
In older versions of mypaint, it was possible to pick a color by simply
holding down the 'pick color' hotkey and hovering over a color. At some
point, this changed. Users were now required to hold down the hotkey and
then drag over the color. This patch restores the old behaviour.

Resolves mypaint/mypaint#253.