Workspace scrolls: fix stacking borders
Seems that we're supposed to use gtk_container_add() in all cases since
GTK 3.8, and not the deprecated gtk_scrolled_window_add_with_viewport().
In fact it looks as if doing the latter was causing a stack of internal
children to surround the canvas widget when MyPaint was repeatedly
fullscreened and unfullscreened.

Closes mypaint/mypaint#92.