Workspace: fix some GTK 3.16+ annoyances
GTK 3.16 and beyond convert a lot of normal UI elements into mystery
meat systems which you have to hover over to discover. Not such a good
thing for our sidebars.

This commit makes the dividers look as though they can be moved again,
and makes the scrollbar visible all the time when the stack's minimal
height is too small to be displayed fully visible (thus preventing it
from overlapping the close button).