Inking tool: force-directed layout for buttons
Use some fake physics for layout to avoid button overlaps and other
messiness. It'll probably need some messing with the constants.

The accept button is placed near the final node to allow simpler
continuation. Hopefully it won't get in the way there.

The reject button is placed near the start node, because if you didn't want
to stroke, you'll presumably be wanting to go back to the beginning and try
again.