Pick Color: remove screen sampling feature
Fix a segfault when picking from the screen by rendering a sample square of
the document under the cursor, and averaging the pixels in that.

This is a removal of an old feature, but the old way of doing things was
never very satisfactory or very well supported across different DEs and
other OS windowing environments. See the long list of bugs about this
feature over time.

Closes mypaint/mypaint#333 - colour pick now ignores display filters.
Closes mypaint/mypaint#444 - picking colours no longer crashes MyPaint.