tdw: suppress redraws during load and save
Prevent TDWs from invalidating areas on screen when the model is
changing while it's loaded or saved. When loading or saving, the UI
is already frozen by putting it in the "insensitive" widget state, so
use that as the controlling flag.

Use the new area-render function to capture a view of the last
sensitive state of the widget minus any overlays, and draw it
while loading or saving for natural invalidations/redraws.

(Turns out we do actually send a full redraw when loading or saving,
so redraw the capture with a bit of alpha to make it look pretty and
behave like widgets in the Adwaita theme, which also dim as of 3.18.x.
Cleverer people than me can work on loading anims if they want...)

Closes mypaint/mypaint#468 by reimplementing a Pixmap-based equivalent
bit of code that was dropped for convenience during the transition from
GTK2 to GTK3.