freehand: ignore alien motions while btn pressed
In lieu of true multi-device stuff (which may not be possible for grab
reasons), respond to button press and button release events by establishing
a filter that causes motion events from other devices to be suppressed
while the button is held down on the device you are painting with. We
assume this latter device is the one that issued the click, just like the
existing workaround for dodgy evdev event sequences.

There may still be categories of input devices that don't send button press
events when the nib is pressed down, but which still send pressureful
motion events. Tablet PC semantics with the side btn1, for example. This
code should remain compatible with those devices.

This may help mypaint/mypaint#460, and certain jaggedy Windows glitches
like mypaint/mypaint#177.