backgroundwindow: do not overwrite background
Removes some old legacy background loading code from the dialog
initialisation, that is not needed any more.
Opening the background chooser dialog for the first time has
overwritten the background of an already open painting with the
default background.
Correct loading of the background is done in gui/application.py.

Solves mypaint/mypaint#512