externalapp: use os.startfile on windows
Use the nonportable os.startfile() way of working on Windows to workaround
a segfault. Users will no longer be offered a choice of what to launch on
that platform: use Explorer to set the default edit association for the
extension you want to edit instead.

Addresses mypaint/mypaint#531 enough to have it not hinder the release, but
this is a pretty silly workaround for an upstream GLib/GDK regression.