Check for use_sharedlib before building lib tests
Prior to this, it would be impossible to run mypaint without the tests
from brushlib. This shouldn't break the Travis build since it doesn't
set use_sharedlib=true. It surrounds the relevant test code with
ifndef's that check for the presence of NO_TESTS passed from g++. If
NO_TESTS is defined, then lib/ avoids building any test related code.