inkmode: fix inter-node steps calculation
Interpolation needs at least two steps. On some Windows configurations, it
was able to calculate zero for the number of needed steps, which is
surprising (and also causes a zero-division error in practice).

For this to happen, the two middle nodes of the interpolation would need to
have been recorded with identical times and positions. Perhaps the
interpolation should just be skipped instead of drawn. The node probably
shouldn't have been recorded either.

Anyway, hopefully fixes mypaint/mypaint#549. But may just move the problem
elsewhere.