1-shot rotate actions: reverse labels & icons
In mypaint/mypaint#553, we decided that one-shot toolbar buttons that
change the view should present themselves as affecting the canvas you see
in the view. It's one less level of abstraction for the user, and thinking
about it the other way only really makes sense for the Preview window.

No string additions, just reordering → translations will be OK.

Removed two tooltips, since their old content would have been confusing
now. I'll raise a new issue about this, because they should be re-added
when we're out of string freeze.

It looks like it was this way originally somehow? The icons look better
this way round anyway.

Closes mypaint/mypaint#553.