lib/SConscript: update brushlib tests LIBPATH
Over in libmypaint, I'm refactoring tests/SConscript to make only a static
PIC library, since the tests shouldn't be built for the new default
shared-lib case (they wouldn't run without libmypaint installation).

When libmypaint is built as the brushlib/ submodule of MyPaint, it is built
statically. We're using the configuration variable to control whether the
tests get built and run now, to give us future CI testing from source but
nice shared-lib releases that don't pull in huge bundles of test crap. Best
of both worlds? I hope so.

Extend LIBPATH so that the mypaintlib Python extension can find
libmypaint-tests.a in the static/testable case.

Refactor the code a bit for clarity.

Update to mypaint/mypaint#542.