Rect: fix cut and paste error in utility ctor
Fix a silly cut and paste error in Rect.new_from_gdk_rectangle(),
and add a test for this code. Test is mocked due to the inability to
construct a real GdkRectangle on some systems (see #437).

This typo was causing mispositioned windows in multi-monitor setups.
Thanks for spotting this, @dothiko!

Addresses mypaint/mypaint#596
May help with mypaint/mypaint#437