Catch+log exceptions when pasting
This can occur when the layer type does not support loading from a bitmap.
Layer groups are one example.

Catchall exception handling is bad, so we'd better add a generic and user-
facing "command or operation failed" exception that can be handled
explicitly for 1.3.0. This will do for 1.2.1 though.

Addresses mypaint/mypaint#603

[Backport of 11d2e7ee6ec3a62a1269f45d9944a484fa98eea7 from master]

This cherry-pick closes mypaint/mypaint#603.