ToolStack: Fix sticking of the last paned.
The problem is that when we create a widget, its parent Gtk.Container
is already connected to the "size-allocate" signal of the parent
Gtk.Widget, so our handlers are always called after all child widgets
have been resized.

The simple solution is to override do_size_allocate() so that our
handler will be called first.
Another good thing is that the separator is now sticks when it's moved
manually.

Fixes mypaint/mypaint#298.