README: use Debian's "new" libpng-dev name
Debian have changed the dev package name with 1.6.21 to not contain a
version code. It's just libpng-dev now, not *libpng16-dev.
We're fine building against the 1.6 libs, so let's do that.

Practically speaking, we're still compatible with both 1.6 and 1.2, but the
dev packages conflict. Users will experience less friction if they start
using the new name due to 2nd order dependencies we don't control.

Formerly libpng-dev was a virtual package, so this name has been around for
a while.

Update the instructions, and mention the earlier libjson0-dev →
libjson-c-dev name migration too. Might as well start listing these.

[Cherry-pick of d91059460a44f00b5eed2931b06cf9f31130ac0e from master]