workspace: tab pages: eat drag-starting events
The page widgets that contain the sidebar panels now fully consume the
types of pointer event that can start a drag. This stops these events
propagating up to the containing GtkNotebook.

Fixes accidental/hypersensitive dragging of the current page's tab when
clicking or dragging in a repositionable notebook page since GTK 3.19.8.

Better fix for mypaint/mypaint#643.

[Cherry-pick of 996a33e14b6bc20cd2ac6af8e1d821f01e1ce697 from master]