adjbases: Revert to previous fallthru handling
Now the problem events are consumed elsewhere, we don't need to
consume them here. But be a bit more explicit about it now.
Sometimes an event really should be consumed in the base class's
internal handler.

Standardize the coding a bit, to make it more obvious.

The earlier fix for mypaint/mypaint#643 was preventing the palette
and brush group menus from triggering ☹