Added a Bug Report Closing Condition
I added a condition to the bottom of the template if the bug reporter dose not fill out the template correctly, the bug report will be closed. This hopefully avoid bug reports like #676 from happening again. Plus we need to draw a hard line from a good bug reports to just plain lazy bug reports.