More issue template tweaks
Bit of parallelism in th' prose.

Move up the warning and tweak. Reporting isn't really a form-filling
exercise, despite the templte's appearance. Any format is OK; trying to
force one standard isn't going to work realistically. But sure: reports
that are going nowhere from the start are free to close.