Generate and upload AppImage of each commit (#740)
Experimental Appimage builds from @probonopd.

Travis will generate and upload an AppImage of each commit. This works by bundling MyPaint, Python, and the required Python modules inside the AppImage.

The URL of the AppImage can be found in the Travis CI build log, beginning with http://transfer.sh/

Please test, and volunteer to maintain this.